refactor(primitives): use alloy's EthereumReceipt type
