+++ crates/ethereum/primitives/Cargo.toml
@@ -15,7 +15,6 @@ workspace = true
 # reth
 reth-codecs = { workspace = true, optional = true }
 reth-primitives-traits.workspace = true
-reth-zstd-compressors = { workspace = true, optional = true }
 
 # ethereum
 alloy-eips = { workspace = true, features = ["k256"] }
@@ -23,22 +22,20 @@ alloy-primitives.workspace = true
 alloy-consensus = { workspace = true, features = ["serde"] }
 alloy-serde = { workspace = true, optional = true }
 alloy-rpc-types-eth = { workspace = true, optional = true }
-alloy-rlp.workspace = true
 
 # misc
 arbitrary = { workspace = true, optional = true, features = ["derive"] }
-modular-bitfield = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 serde_with = { workspace = true, optional = true }
 
 [dev-dependencies]
 derive_more.workspace = true
 arbitrary.workspace = true
-bincode.workspace = true
 proptest.workspace = true
 proptest-arbitrary-interop.workspace = true
 rand_08.workspace = true
 rand.workspace = true
+alloy-rlp.workspace = true
 reth-codecs = { workspace = true, features = ["test-utils"] }
 reth-zstd-compressors.workspace = true
 secp256k1 = { workspace = true, features = ["rand"] }
@@ -54,9 +51,7 @@ test-utils = [
 std = [
     "alloy-consensus/std",
     "alloy-primitives/std",
-    "alloy-rlp/std",
     "reth-primitives-traits/std",
-    "reth-zstd-compressors?/std",
     "serde?/std",
     "alloy-eips/std",
     "derive_more/std",
@@ -69,8 +64,6 @@ std = [
 reth-codec = [
     "std",
     "dep:reth-codecs",
-    "dep:modular-bitfield",
-    "dep:reth-zstd-compressors",
 ]
 arbitrary = [
     "std",
