+++ crates/ethereum/primitives/src/receipt.rs
@@ -1,42 +1,8 @@
-use core::fmt::Debug;
-
-use alloc::vec::Vec;
-use alloy_consensus::{
-    Eip2718DecodableReceipt, Eip2718EncodableReceipt, Eip658Value, ReceiptEnvelope,
-    ReceiptWithBloom, RlpDecodableReceipt, RlpEncodableReceipt, TxReceipt, TxType, Typed2718,
-};
-use alloy_eips::eip2718::{Eip2718Error, Eip2718Result, Encodable2718, IsTyped2718};
-use alloy_primitives::{Bloom, Log, B256};
-use alloy_rlp::{BufMut, Decodable, Encodable, Header, RlpDecodable, RlpEncodable};
-use reth_primitives_traits::{proofs::ordered_trie_root_with_encoder, InMemorySize};
-
-/// Helper trait alias with requirements for transaction type generic to be used within [`Receipt`].
-pub trait TxTy:
-    Debug
-    + Copy
-    + Eq
-    + Send
-    + Sync
-    + InMemorySize
-    + Typed2718
-    + TryFrom<u8, Error = Eip2718Error>
-    + Decodable
-    + 'static
-{
-}
-impl<T> TxTy for T where
-    T: Debug
-        + Copy
-        + Eq
-        + Send
-        + Sync
-        + InMemorySize
-        + Typed2718
-        + TryFrom<u8, Error = Eip2718Error>
-        + Decodable
-        + 'static
-{
-}
+use alloy_consensus::TxType;
+pub use alloy_consensus::{EthereumReceipt, TxTy};
+use alloy_eips::eip2718::Encodable2718;
+use alloy_primitives::B256;
+use reth_primitives_traits::proofs::ordered_trie_root_with_encoder;
 
 /// Raw ethereum receipt.
 pub type Receipt<T = TxType> = EthereumReceipt<T>;
@@ -45,33 +11,20 @@ pub type Receipt<T = TxType> = EthereumReceipt<T>;
 /// Receipt representation for RPC.
 pub type RpcReceipt<T = TxType> = EthereumReceipt<T, alloy_rpc_types_eth::Log>;
 
-/// Typed ethereum transaction receipt.
-/// Receipt containing result of transaction execution.
-#[derive(Clone, Debug, PartialEq, Eq, Default, RlpEncodable, RlpDecodable)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
-#[cfg_attr(feature = "reth-codec", reth_codecs::add_arbitrary_tests(compact, rlp))]
-#[cfg_attr(feature = "serde", serde(rename_all = "camelCase"))]
-pub struct EthereumReceipt<T = TxType, L = Log> {
-    /// Receipt type.
-    #[cfg_attr(feature = "serde", serde(rename = "type"))]
-    pub tx_type: T,
-    /// If transaction is executed successfully.
-    ///
-    /// This is the `statusCode`
-    #[cfg_attr(feature = "serde", serde(with = "alloy_serde::quantity", rename = "status"))]
-    pub success: bool,
-    /// Gas used
-    #[cfg_attr(feature = "serde", serde(with = "alloy_serde::quantity"))]
-    pub cumulative_gas_used: u64,
-    /// Log send from contracts.
-    pub logs: Vec<L>,
+/// Extension trait for [`Receipt`] providing reth-specific methods.
+pub trait ReceiptExt<T> {
+    /// Converts the logs of the receipt to RPC logs.
+    #[cfg(feature = "rpc")]
+    fn into_rpc(
+        self,
+        next_log_index: usize,
+        meta: alloy_consensus::transaction::TransactionMeta,
+    ) -> RpcReceipt<T>;
 }
 
 #[cfg(feature = "rpc")]
-impl<T> Receipt<T> {
-    /// Converts the logs of the receipt to RPC logs.
-    pub fn into_rpc(
+impl<T> ReceiptExt<T> for Receipt<T> {
+    fn into_rpc(
         self,
         next_log_index: usize,
         meta: alloy_consensus::transaction::TransactionMeta,
@@ -82,491 +35,25 @@ impl<T> Receipt<T> {
     }
 }
 
-impl<T: TxTy> Receipt<T> {
-    /// Returns length of RLP-encoded receipt fields with the given [`Bloom`] without an RLP header.
-    pub fn rlp_encoded_fields_length(&self, bloom: &Bloom) -> usize {
-        self.success.length() +
-            self.cumulative_gas_used.length() +
-            bloom.length() +
-            self.logs.length()
-    }
-
-    /// RLP-encodes receipt fields with the given [`Bloom`] without an RLP header.
-    pub fn rlp_encode_fields(&self, bloom: &Bloom, out: &mut dyn BufMut) {
-        self.success.encode(out);
-        self.cumulative_gas_used.encode(out);
-        bloom.encode(out);
-        self.logs.encode(out);
-    }
-
-    /// Returns RLP header for inner encoding.
-    pub fn rlp_header_inner(&self, bloom: &Bloom) -> Header {
-        Header { list: true, payload_length: self.rlp_encoded_fields_length(bloom) }
-    }
-
-    /// RLP-decodes the receipt from the provided buffer. This does not expect a type byte or
-    /// network header.
-    pub fn rlp_decode_inner(
-        buf: &mut &[u8],
-        tx_type: T,
-    ) -> alloy_rlp::Result<ReceiptWithBloom<Self>> {
-        let header = Header::decode(buf)?;
-        if !header.list {
-            return Err(alloy_rlp::Error::UnexpectedString);
-        }
-
-        let remaining = buf.len();
-
-        let success = Decodable::decode(buf)?;
-        let cumulative_gas_used = Decodable::decode(buf)?;
-        let logs_bloom = Decodable::decode(buf)?;
-        let logs = Decodable::decode(buf)?;
-
-        if buf.len() + header.payload_length != remaining {
-            return Err(alloy_rlp::Error::UnexpectedLength);
-        }
-
-        Ok(ReceiptWithBloom {
-            receipt: Self { cumulative_gas_used, tx_type, success, logs },
-            logs_bloom,
-        })
-    }
-
-    /// Calculates the receipt root for a header for the reference type of [Receipt].
-    ///
-    /// NOTE: Prefer `proofs::calculate_receipt_root` if you have log blooms memoized.
-    pub fn calculate_receipt_root_no_memo(receipts: &[Self]) -> B256 {
-        ordered_trie_root_with_encoder(receipts, |r, buf| r.with_bloom_ref().encode_2718(buf))
-    }
+/// Calculates the receipt root for a header for the reference type of [`Receipt`].
+///
+/// NOTE: Prefer `proofs::calculate_receipt_root` if you have log blooms memoized.
+pub fn calculate_receipt_root_no_memo<T: TxTy>(receipts: &[Receipt<T>]) -> B256 {
+    ordered_trie_root_with_encoder(receipts, |r, buf| {
+        alloy_consensus::TxReceipt::with_bloom_ref(r).encode_2718(buf)
+    })
 }
-
-impl<T: TxTy> Eip2718EncodableReceipt for Receipt<T> {
-    fn eip2718_encoded_length_with_bloom(&self, bloom: &Bloom) -> usize {
-        !self.tx_type.is_legacy() as usize + self.rlp_header_inner(bloom).length_with_payload()
-    }
-
-    fn eip2718_encode_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut) {
-        if !self.tx_type.is_legacy() {
-            out.put_u8(self.tx_type.ty());
-        }
-        self.rlp_header_inner(bloom).encode(out);
-        self.rlp_encode_fields(bloom, out);
-    }
-}
-
-impl<T: TxTy> Eip2718DecodableReceipt for Receipt<T> {
-    fn typed_decode_with_bloom(ty: u8, buf: &mut &[u8]) -> Eip2718Result<ReceiptWithBloom<Self>> {
-        Ok(Self::rlp_decode_inner(buf, T::try_from(ty)?)?)
-    }
-
-    fn fallback_decode_with_bloom(buf: &mut &[u8]) -> Eip2718Result<ReceiptWithBloom<Self>> {
-        Ok(Self::rlp_decode_inner(buf, T::try_from(0)?)?)
-    }
-}
-
-impl<T: TxTy> RlpEncodableReceipt for Receipt<T> {
-    fn rlp_encoded_length_with_bloom(&self, bloom: &Bloom) -> usize {
-        let mut len = self.eip2718_encoded_length_with_bloom(bloom);
-        if !self.tx_type.is_legacy() {
-            len += Header {
-                list: false,
-                payload_length: self.eip2718_encoded_length_with_bloom(bloom),
-            }
-            .length();
-        }
-
-        len
-    }
-
-    fn rlp_encode_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut) {
-        if !self.tx_type.is_legacy() {
-            Header { list: false, payload_length: self.eip2718_encoded_length_with_bloom(bloom) }
-                .encode(out);
-        }
-        self.eip2718_encode_with_bloom(bloom, out);
-    }
-}
-
-impl<T: TxTy> RlpDecodableReceipt for Receipt<T> {
-    fn rlp_decode_with_bloom(buf: &mut &[u8]) -> alloy_rlp::Result<ReceiptWithBloom<Self>> {
-        let header_buf = &mut &**buf;
-        let header = Header::decode(header_buf)?;
-
-        // Legacy receipt, reuse initial buffer without advancing
-        if header.list {
-            return Self::rlp_decode_inner(buf, T::try_from(0)?)
-        }
-
-        // Otherwise, advance the buffer and try decoding type flag followed by receipt
-        *buf = *header_buf;
-
-        let remaining = buf.len();
-        let tx_type = T::decode(buf)?;
-        let this = Self::rlp_decode_inner(buf, tx_type)?;
-
-        if buf.len() + header.payload_length != remaining {
-            return Err(alloy_rlp::Error::UnexpectedLength);
-        }
-
-        Ok(this)
-    }
-}
-
-impl<T, L> TxReceipt for EthereumReceipt<T, L>
-where
-    T: TxTy,
-    L: Send + Sync + Clone + Debug + Eq + AsRef<Log>,
-{
-    type Log = L;
-
-    fn status_or_post_state(&self) -> Eip658Value {
-        self.success.into()
-    }
-
-    fn status(&self) -> bool {
-        self.success
-    }
-
-    fn bloom(&self) -> Bloom {
-        alloy_primitives::logs_bloom(self.logs.iter().map(|l| l.as_ref()))
-    }
-
-    fn cumulative_gas_used(&self) -> u64 {
-        self.cumulative_gas_used
-    }
-
-    fn logs(&self) -> &[L] {
-        &self.logs
-    }
-
-    fn into_logs(self) -> Vec<L> {
-        self.logs
-    }
-}
-
-impl<T: TxTy> Typed2718 for Receipt<T> {
-    fn ty(&self) -> u8 {
-        self.tx_type.ty()
-    }
-}
-
-impl<T: TxTy> IsTyped2718 for Receipt<T> {
-    fn is_type(type_id: u8) -> bool {
-        <TxType as IsTyped2718>::is_type(type_id)
-    }
-}
-
-impl<T: TxTy> InMemorySize for Receipt<T> {
-    fn size(&self) -> usize {
-        size_of::<Self>() + self.logs.iter().map(|log| log.size()).sum::<usize>()
-    }
-}
-
-impl<T> From<ReceiptEnvelope<T>> for Receipt<TxType>
-where
-    T: Into<Log>,
-{
-    fn from(value: ReceiptEnvelope<T>) -> Self {
-        let value = value.into_primitives_receipt();
-        Self {
-            tx_type: value.tx_type(),
-            success: value.is_success(),
-            cumulative_gas_used: value.cumulative_gas_used(),
-            logs: value.into_logs(),
-        }
-    }
-}
-
-impl<T, L> From<EthereumReceipt<T, L>> for alloy_consensus::Receipt<L> {
-    fn from(value: EthereumReceipt<T, L>) -> Self {
-        Self {
-            status: value.success.into(),
-            cumulative_gas_used: value.cumulative_gas_used,
-            logs: value.logs,
-        }
-    }
-}
-
-impl<L> From<EthereumReceipt<TxType, L>> for ReceiptEnvelope<L>
-where
-    L: Send + Sync + Clone + Debug + Eq + AsRef<Log>,
-{
-    fn from(value: EthereumReceipt<TxType, L>) -> Self {
-        let tx_type = value.tx_type;
-        let receipt = value.into_with_bloom().map_receipt(Into::into);
-        match tx_type {
-            TxType::Legacy => Self::Legacy(receipt),
-            TxType::Eip2930 => Self::Eip2930(receipt),
-            TxType::Eip1559 => Self::Eip1559(receipt),
-            TxType::Eip4844 => Self::Eip4844(receipt),
-            TxType::Eip7702 => Self::Eip7702(receipt),
-        }
-    }
-}
-
-#[cfg(all(feature = "serde", feature = "serde-bincode-compat"))]
-pub(super) mod serde_bincode_compat {
-    use alloc::{borrow::Cow, vec::Vec};
-    use alloy_consensus::TxType;
-    use alloy_eips::eip2718::Eip2718Error;
-    use alloy_primitives::{Log, U8};
-    use core::fmt::Debug;
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-    use serde_with::{DeserializeAs, SerializeAs};
-
-    /// Bincode-compatible [`super::Receipt`] serde implementation.
-    ///
-    /// Intended to use with the [`serde_with::serde_as`] macro in the following way:
-    /// ```rust
-    /// use alloy_consensus::TxType;
-    /// use reth_ethereum_primitives::{serde_bincode_compat, Receipt};
-    /// use serde::{de::DeserializeOwned, Deserialize, Serialize};
-    /// use serde_with::serde_as;
-    ///
-    /// #[serde_as]
-    /// #[derive(Serialize, Deserialize)]
-    /// struct Data {
-    ///     #[serde_as(as = "serde_bincode_compat::Receipt<'_>")]
-    ///     receipt: Receipt<TxType>,
-    /// }
-    /// ```
-    #[derive(Debug, Serialize, Deserialize)]
-    #[serde(bound(deserialize = "T: TryFrom<u8, Error = Eip2718Error>"))]
-    pub struct Receipt<'a, T = TxType> {
-        /// Receipt type.
-        #[serde(deserialize_with = "deserde_txtype")]
-        pub tx_type: T,
-        /// If transaction is executed successfully.
-        ///
-        /// This is the `statusCode`
-        pub success: bool,
-        /// Gas used
-        pub cumulative_gas_used: u64,
-        /// Log send from contracts.
-        pub logs: Cow<'a, Vec<Log>>,
-    }
-
-    /// Ensures that txtype is deserialized symmetrically as U8
-    fn deserde_txtype<'de, D, T>(deserializer: D) -> Result<T, D::Error>
-    where
-        D: Deserializer<'de>,
-        T: TryFrom<u8, Error = Eip2718Error>,
-    {
-        U8::deserialize(deserializer)?.to::<u8>().try_into().map_err(serde::de::Error::custom)
-    }
-
-    impl<'a, T: Copy> From<&'a super::Receipt<T>> for Receipt<'a, T> {
-        fn from(value: &'a super::Receipt<T>) -> Self {
-            Self {
-                tx_type: value.tx_type,
-                success: value.success,
-                cumulative_gas_used: value.cumulative_gas_used,
-                logs: Cow::Borrowed(&value.logs),
-            }
-        }
-    }
-
-    impl<'a, T> From<Receipt<'a, T>> for super::Receipt<T> {
-        fn from(value: Receipt<'a, T>) -> Self {
-            Self {
-                tx_type: value.tx_type,
-                success: value.success,
-                cumulative_gas_used: value.cumulative_gas_used,
-                logs: value.logs.into_owned(),
-            }
-        }
-    }
-
-    impl<T: Copy + Serialize> SerializeAs<super::Receipt<T>> for Receipt<'_, T> {
-        fn serialize_as<S>(source: &super::Receipt<T>, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            Receipt::<'_>::from(source).serialize(serializer)
-        }
-    }
-
-    impl<'de, T: TryFrom<u8, Error = Eip2718Error>> DeserializeAs<'de, super::Receipt<T>>
-        for Receipt<'de, T>
-    {
-        fn deserialize_as<D>(deserializer: D) -> Result<super::Receipt<T>, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            Receipt::<'_, T>::deserialize(deserializer).map(Into::into)
-        }
-    }
-
-    impl<T> reth_primitives_traits::serde_bincode_compat::SerdeBincodeCompat for super::Receipt<T>
-    where
-        T: Copy + Serialize + TryFrom<u8, Error = Eip2718Error> + Debug + 'static,
-    {
-        type BincodeRepr<'a> = Receipt<'a, T>;
-
-        fn as_repr(&self) -> Self::BincodeRepr<'_> {
-            self.into()
-        }
-
-        fn from_repr(repr: Self::BincodeRepr<'_>) -> Self {
-            repr.into()
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use crate::{receipt::serde_bincode_compat, Receipt};
-        use alloy_consensus::TxType;
-        use arbitrary::Arbitrary;
-        use rand::Rng;
-        use serde_with::serde_as;
-
-        #[test]
-        fn test_receipt_bincode_roundtrip() {
-            #[serde_as]
-            #[derive(Debug, PartialEq, Eq)]
-            #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-            struct Data {
-                #[serde_as(as = "serde_bincode_compat::Receipt<'_, TxType>")]
-                receipt: Receipt<TxType>,
-            }
-
-            let mut bytes = [0u8; 1024];
-            rand::rng().fill(bytes.as_mut_slice());
-            let data = Data {
-                receipt: Receipt::arbitrary(&mut arbitrary::Unstructured::new(&bytes)).unwrap(),
-            };
-            let encoded = bincode::serialize(&data).unwrap();
-            let decoded: Data = bincode::deserialize(&encoded).unwrap();
-            assert_eq!(decoded, data);
-        }
-    }
-}
-
-#[cfg(feature = "reth-codec")]
-mod compact {
-    use super::*;
-    use reth_codecs::{
-        Compact,
-        __private::{modular_bitfield::prelude::*, Buf},
-    };
-
-    impl Receipt {
-        #[doc = "Used bytes by [`ReceiptFlags`]"]
-        pub const fn bitflag_encoded_bytes() -> usize {
-            1u8 as usize
-        }
-        #[doc = "Unused bits for new fields by [`ReceiptFlags`]"]
-        pub const fn bitflag_unused_bits() -> usize {
-            0u8 as usize
-        }
-    }
-
-    #[allow(non_snake_case, unused_parens)]
-    mod flags {
-        use super::*;
-
-        #[doc = "Fieldset that facilitates compacting the parent type. Used bytes: 1 | Unused bits: 0"]
-        #[bitfield]
-        #[derive(Clone, Copy, Debug, Default)]
-        pub struct ReceiptFlags {
-            pub tx_type_len: B2,
-            pub success_len: B1,
-            pub cumulative_gas_used_len: B4,
-            pub __zstd: B1,
-        }
-
-        impl ReceiptFlags {
-            #[doc = r" Deserializes this fieldset and returns it, alongside the original slice in an advanced position."]
-            pub fn from(mut buf: &[u8]) -> (Self, &[u8]) {
-                (Self::from_bytes([buf.get_u8()]), buf)
-            }
-        }
-    }
-
-    pub use flags::ReceiptFlags;
-
-    impl<T: Compact> Compact for Receipt<T> {
-        fn to_compact<B>(&self, buf: &mut B) -> usize
-        where
-            B: reth_codecs::__private::bytes::BufMut + AsMut<[u8]>,
-        {
-            let mut flags = ReceiptFlags::default();
-            let mut total_length = 0;
-            let mut buffer = reth_codecs::__private::bytes::BytesMut::new();
-
-            let tx_type_len = self.tx_type.to_compact(&mut buffer);
-            flags.set_tx_type_len(tx_type_len as u8);
-            let success_len = self.success.to_compact(&mut buffer);
-            flags.set_success_len(success_len as u8);
-            let cumulative_gas_used_len = self.cumulative_gas_used.to_compact(&mut buffer);
-            flags.set_cumulative_gas_used_len(cumulative_gas_used_len as u8);
-            self.logs.to_compact(&mut buffer);
-
-            let zstd = buffer.len() > 7;
-            if zstd {
-                flags.set___zstd(1);
-            }
-
-            let flags = flags.into_bytes();
-            total_length += flags.len() + buffer.len();
-            buf.put_slice(&flags);
-            if zstd {
-                reth_zstd_compressors::with_receipt_compressor(|compressor| {
-                    let compressed = compressor.compress(&buffer).expect("Failed to compress.");
-                    buf.put(compressed.as_slice());
-                });
-            } else {
-                buf.put(buffer);
-            }
-            total_length
-        }
-
-        fn from_compact(buf: &[u8], _len: usize) -> (Self, &[u8]) {
-            let (flags, mut buf) = ReceiptFlags::from(buf);
-            if flags.__zstd() != 0 {
-                reth_zstd_compressors::with_receipt_decompressor(|decompressor| {
-                    let decompressed = decompressor.decompress(buf);
-                    let original_buf = buf;
-                    let mut buf: &[u8] = decompressed;
-                    let (tx_type, new_buf) = T::from_compact(buf, flags.tx_type_len() as usize);
-                    buf = new_buf;
-                    let (success, new_buf) = bool::from_compact(buf, flags.success_len() as usize);
-                    buf = new_buf;
-                    let (cumulative_gas_used, new_buf) =
-                        u64::from_compact(buf, flags.cumulative_gas_used_len() as usize);
-                    buf = new_buf;
-                    let (logs, _) = Vec::from_compact(buf, buf.len());
-                    (Self { tx_type, success, cumulative_gas_used, logs }, original_buf)
-                })
-            } else {
-                let (tx_type, new_buf) = T::from_compact(buf, flags.tx_type_len() as usize);
-                buf = new_buf;
-                let (success, new_buf) = bool::from_compact(buf, flags.success_len() as usize);
-                buf = new_buf;
-                let (cumulative_gas_used, new_buf) =
-                    u64::from_compact(buf, flags.cumulative_gas_used_len() as usize);
-                buf = new_buf;
-                let (logs, new_buf) = Vec::from_compact(buf, buf.len());
-                buf = new_buf;
-                let obj = Self { tx_type, success, cumulative_gas_used, logs };
-                (obj, buf)
-            }
-        }
-    }
-}
-
-#[cfg(feature = "reth-codec")]
-pub use compact::*;
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::TransactionSigned;
+    use alloy_consensus::{ReceiptWithBloom, TxReceipt, TxType};
     use alloy_eips::eip2718::Encodable2718;
     use alloy_primitives::{
-        address, b256, bloom, bytes, hex_literal::hex, Address, Bytes, Log, LogData,
+        address, b256, bloom, bytes, hex_literal::hex, Address, Bloom, Bytes, Log, LogData,
     };
-    use alloy_rlp::Decodable;
+    use alloy_rlp::{Decodable, Encodable};
     use reth_codecs::Compact;
     use reth_primitives_traits::proofs::{
         calculate_receipt_root, calculate_transaction_root, calculate_withdrawals_root,
@@ -784,6 +271,8 @@ mod tests {
     #[test]
     #[cfg(feature = "rpc")]
     fn test_receipt_serde() {
+        use alloy_consensus::ReceiptEnvelope;
+
         let input = r#"{"status":"0x1","cumulativeGasUsed":"0x175cc0e","logs":[{"address":"0xa18b9ca2a78660d44ab38ae72e72b18792ffe413","topics":["0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925","0x000000000000000000000000e7e7d8006cbff47bc6ac2dabf592c98e97502708","0x0000000000000000000000007a250d5630b4cf539739df2c5dacb4c659f2488d"],"data":"0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff","blockHash":"0xbf9e6a368a399f996a0f0b27cab4191c028c3c99f5f76ea08a5b70b961475fcb","blockNumber":"0x164b59f","blockTimestamp":"0x68c9a713","transactionHash":"0x533aa9e57865675bb94f41aa2895c0ac81eee69686c77af16149c301e19805f1","transactionIndex":"0x14d","logIndex":"0x238","removed":false}],"logsBloom":"0x00000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000400000040000000000000004000000000000000000000000000000000000000000000020000000000000000000000000080000000000000000000000000200000020000000000000000000000000000000000000000000000000000000000000020000010000000000000000000000000000000000000000000000000000000000000","type":"0x2","transactionHash":"0x533aa9e57865675bb94f41aa2895c0ac81eee69686c77af16149c301e19805f1","transactionIndex":"0x14d","blockHash":"0xbf9e6a368a399f996a0f0b27cab4191c028c3c99f5f76ea08a5b70b961475fcb","blockNumber":"0x164b59f","gasUsed":"0xb607","effectiveGasPrice":"0x4a3ee768","from":"0xe7e7d8006cbff47bc6ac2dabf592c98e97502708","to":"0xa18b9ca2a78660d44ab38ae72e72b18792ffe413","contractAddress":null}"#;
         let receipt: RpcReceipt = serde_json::from_str(input).unwrap();
         let envelope: ReceiptEnvelope<alloy_rpc_types_eth::Log> =
