+++ crates/storage/codecs/src/alloy/mod.rs
@@ -18,6 +18,7 @@ cond_mod!(
     genesis_account,
     header,
     log,
+    receipt,
     signature,
     trie,
     txkind,
